Updated the gem dependencies and installer.
The gem now requires all of the "agnostic" and "english" gems that
the installed used to install, and the installer now checks for
installed gems before trying to install them itself. This way, bundler
will install these dependencies for us.

Also, fixed some path issues with the installer so that the model
files, etc., all go where they're supposed to.